+++ src/store/reducers/metadataReducer.js
@@ -8,8 +8,7 @@ export default function (storage) {
     linkTarget: '',
     domHighlight: Map(),
     hintText: '',
-    showTooltip: false,
-    context: ''
+    showTooltip: false
   });
 
   const initialState = Map({
@@ -45,9 +44,6 @@ export default function (storage) {
       }
       case actionTypes.SET_HINT_TEXT: {
         return storeMetadata(state.set('hintText', action.hint));
-      }
-      case actionTypes.SET_CONTEXT: {
-        return storeMetadata(state.set('context', action.context));
       }
       case actionTypes.PULL_SESSION: {
         const localSession = getLocalSession(storage, SESSION_NAME);